+++ func_to_web/__init__.py
@@ -4,5 +4,5 @@ from typing import Annotated, Literal
 from datetime import date, time
 from pydantic import Field
 
-__version__ = "0.9.7"
+__version__ = "0.9.8"
 __all__ = ["run", "Annotated", "Literal", "date", "time", "Field"]